+++ include/bitmap.h
@@ -13,25 +13,26 @@
 #define INCLUDE_BITMAP_H_
 
 #include "stdint.h"
+#include "stddef.h"
 
 typedef struct {
 	uint8_t *buffer;
-	unsigned long length;
+	size_t length;
 } Bitmap;
 
 /* Initialize the memory bitmap */
-void bitmap_init(Bitmap *bitmap, uint8_t *buffer, unsigned long size);
+void bitmap_init(Bitmap *bitmap, uint8_t *buffer, size_t size);
 
 /* Get memory bitmap */
-int bitmap_get(const Bitmap *bitmap, unsigned long index);
+int bitmap_get(const Bitmap *bitmap, size_t index);
 
 /* Setting the memory bitmap */
-void bitmap_set(Bitmap *bitmap, unsigned long index, int value);
+void bitmap_set(Bitmap *bitmap, size_t index, int value);
 
 /* Set the memory bitmap range */
-void bitmap_set_range(Bitmap *bitmap, unsigned long start, unsigned long end, int value);
+void bitmap_set_range(Bitmap *bitmap, size_t start, size_t end, int value);
 
 /* Memory bitmap search range */
-unsigned long bitmap_find_range(const Bitmap *bitmap, unsigned long length, int value);
+size_t bitmap_find_range(const Bitmap *bitmap, size_t length, int value);
 
 #endif // INCLUDE_BITMAP_H_
