+++ libs/string.c
@@ -10,9 +10,11 @@
  */
 
 #include "string.h"
+#include "stdint.h"
+#include "stddef.h"
 
 /* Copy n bytes from memory area str2 to memory area str1 */
-inline void *memcpy(void *str1, const void *str2, unsigned long n)
+inline void *memcpy(void *str1, const void *str2, size_t n)
 {
 	__volatile__ uint8_t *dest = (__volatile__ uint8_t *)str1;
 	__volatile__ const uint8_t *src = (__volatile__ const uint8_t *)str2;
@@ -24,7 +26,7 @@ inline void *memcpy(void *str1, const void *str2, unsigned long n)
 }
 
 /* Sets a memory area to the specified value */
-inline void *memset(void *str, int c, unsigned long n)
+inline void *memset(void *str, int c, size_t n)
 {
 	__volatile__ uint8_t *_str = (__volatile__ uint8_t *)str;
 	__volatile__ uint8_t *end = (__volatile__ uint8_t *)((uint8_t *)str + n);
@@ -35,7 +37,7 @@ inline void *memset(void *str, int c, unsigned long n)
 }
 
 /* Copies n characters from str2 to str1, accounting for overlaps */
-inline void *memmove(void *str1, const void *str2, unsigned long n)
+inline void *memmove(void *str1, const void *str2, size_t n)
 {
 	__volatile__ uint8_t *dest = (__volatile__ uint8_t *)str1;
 	__volatile__ const uint8_t *src = (__volatile__ const uint8_t *)str2;
@@ -52,7 +54,7 @@ inline void *memmove(void *str1, const void *str2, unsigned long n)
 }
 
 /* Compares the first n bytes of memory area str1 with those of memory area str2 */
-inline int memcmp(const void *str1, const void *str2, unsigned long n)
+inline int memcmp(const void *str1, const void *str2, size_t n)
 {
 	const uint8_t *_str1 = (const uint8_t *)str1;
 	const uint8_t *_str2 = (const uint8_t *)str2;
@@ -66,9 +68,9 @@ inline int memcmp(const void *str1, const void *str2, unsigned long n)
 }
 
 /* Calculates the length of the string str */
-inline unsigned long strlen(const char *str)
+inline size_t strlen(const char *str)
 {
-	unsigned long len = 0;
+	size_t len = 0;
 	while (*str++ != '\0') len++;
 	return len;
 }
@@ -82,7 +84,7 @@ inline char *strcpy(char *dest, const char *src)
 }
 
 /* Copies the string pointed to by src to dest, up to n characters. */
-inline char *strncpy(char *dest, const char *src, unsigned long n)
+inline char *strncpy(char *dest, const char *src, size_t n)
 {
 	char *_dest = dest;
 	const char *end = src + n;
@@ -106,7 +108,7 @@ inline int strcmp(const char *str1, const char *str2)
 }
 
 /* Compares the first n characters of two strings for equality */
-inline int strncmp(const char *str1, const char *str2, unsigned long n)
+inline int strncmp(const char *str1, const char *str2, size_t n)
 {
 	const uint8_t *_str1 = (const uint8_t *)str1;
 	const uint8_t *end = (const uint8_t *)str1 + n;
@@ -154,14 +156,14 @@ inline char *strrchr(const char *str, int c)
 /* Find the first occurrence of the string needle in the string haystack, excluding the terminator */
 inline char *strstr(const char *haystack, const char *needle)
 {
-	unsigned long _sn = strlen(haystack), _tn = strlen(needle);
+	size_t _sn = strlen(haystack), _tn = strlen(needle);
 
 	if (_tn == 0) return (char *)haystack;
 	if (_sn < _tn) return 0;
 
 	const char *s = haystack, *t = needle;
 
-	for (unsigned long i = 0; i <= _sn - _tn; i++) {
+	for (size_t i = 0; i <= _sn - _tn; i++) {
 		if (strncmp(s + i, t, _tn) == 0) return (char *)(s + i);
 	}
 	return 0;
